Refactoring LikelihoodOfSearch
