+++ traffic_stops/static/js/app/states/md/LikelihoodOfSearch.js
@@ -17,139 +17,28 @@ var LikelihoodSearchHandler = C.LikelihoodSearchHandlerBase.extend({
   defaults: Stops
 });
 
-var LikelihoodOfSearch = VisualBase.extend({
+var LikelihoodOfSearch = C.LikelihoodOfSearchBase.extend({
   defaults: {
     showEthnicity: true,
     width: 750,
-    height: 375
+    height: 375,
+    Stops: Stops
   },
 
-  setDefaultChart: function () {
-    this.chart = nv.models.multiBarHorizontalChart()
-      .x((d) => d.label)
-      .y((d) => d.value)
-      .width(this.get("width"))
-      .height(this.get("height"))
-      .margin({top: 20, right: 50, bottom: 20, left: 180})
-      .showValues(true)
-      .tooltips(true)
-      .transitionDuration(350)
-      .showControls(false);
-
-    this.chart.yAxis
-        .axisLabel('Additional percentage or search by search-cause')
-        .tickFormat(d3.format('%'));
-
-    this.chart.valueFormat(d3.format('%'));
+  _items: function () {
+    return Stops.ethnicities;
   },
 
-  drawStartup: function () {
-    // get year options for pulldown menu
-    var selector = $('<select>'),
-        year_options = this.data.years,
-        opts = year_options.map((v) => `<option value="${v}">${v}</option>`),
-        getData = () => {
-          var year = selector.val();
-          year = parseInt(year, 10) || year;
-          this.dataset =  this._getDataset(year);
-          this.drawChart();
-        };
-
-    selector
-      .append(opts)
-      .val("Total")
-      .on('change', getData)
-      .trigger('change');
-
-    $('<div>')
-      .html(selector)
-      .appendTo(this.div);
-
-    this.selector = selector;
+  _base: function () {
+    return "White";
   },
 
-  drawChart: function () {
-
-    d3.select(this.svg[0])
-      .datum(this.dataset)
-      .attr('width', "100%")
-      .attr('height', "100%")
-      .attr('preserveAspectRatio', "xMinYMin")
-      .attr('viewBox', `0 0 ${this.get('width')} ${this.get('height')}`)
-      .call(this.chart);
-
-    nv.utils.windowResize(this.chart.update);
+  _defRace: function () {
+    return "Black";
   },
 
-  _getDataset: function (year) {
-    var dataset = [],
-        raw = this.data.raw,
-        stops_arr = raw.stops.filter((v) => v.year===year),
-        searches_arr = raw.searches.filter((v) => v.year===year),
-        stops = d3.map(),
-        searches = d3.map(),
-        items = Stops.ethnicities,
-        base = "White",
-        defRace = "Black",
-        baseUpper = ((d) => d.charAt(0).toUpperCase() + d.slice(1))(base);
-
-    // turn arrays into maps with purpose as the key
-    stops_arr.forEach((v) => {
-      stops.set(v.purpose, v);
-    });
-    searches_arr.forEach((v) => {
-      searches.set(v.purpose, v);
-    });
-
-    // build a set of bars for each ethnicity, except for base
-    items.forEach((ethnicity, i) => {
-      if (ethnicity === base) return;
-
-      var bar = {
-          color: Stops.colors[i],
-          key: `${ethnicity} vs. ${baseUpper}`,
-          values: [],
-          disabled: (ethnicity !== defRace)
-      };
-
-      // build a bar for each violation
-      Stops.purpose_order.forEach((purpose) => {
-        // optional reporting requirement; remove as it's generally unreported
-        if (purpose === "Checkpoint") return;
-
-        // calculate percent-difference of stops which led to searches by ethnicity,
-        // in comparison to base-baseline
-        var search = searches.get(purpose),
-            stop  = stops.get(purpose);
-
-        if (search && stop){
-
-          var rate, base_rate, r_rate,
-              base_se = search[base] || 0,
-              base_st = stop[base] || 0,
-              r_se = search[ethnicity] || 0,
-              r_st = stop[ethnicity] || 0;
-
-          base_rate = base_se/base_st;
-          r_rate = r_se/r_st;
-          rate = (r_rate-base_rate)/base_rate;
-          if(!r_rate || !isFinite(rate)) rate = 0;
-
-          // add purpose to list of values
-          bar.values.push({
-            label: purpose,
-            value: rate,
-            order: Stops.purpose_order.get(purpose)
-          });
-        }
-      });
-
-      // sort bars and then push ethnicity to list
-      bar.values.sort(function(a,b){return a.order-b.order;});
-      dataset.push(bar);
-    });
-
-    return dataset;
+  _pprint: function (type) {
+    return type;
   },
 
   triggerRaceToggle: () => null
